(maint) Refactor lexer specific components to lexer
Previously many of the lexer concerns were located in the parser package and
were marked private. In particular the actual Token IDs and the use of parser
errors in the String Reader.

This commit moves as much of the lexer information into a new lexer package.
The default lexer can not be moved as the lexer calls into the parser for parts
of the logic, which then causes a circular import loop.